+++ tests/config.py
@@ -47,7 +47,7 @@ class Config_:
         assert c.connect_kwargs == {}
         assert c.timeouts.connect is None
         assert c.ssh_config_path is None
-        assert c.inline_ssh_env is False
+        assert c.inline_ssh_env is True
 
     def overrides_some_Invoke_defaults(self):
         config = Config()